Chatbot/requirements.txt
streamlit
streamlit-authenticator
langchain
ollama
pytesseract
pillow
chromadb
pyttsx3
whisper
sounddevice
wavio
